Nasty times...
In the name of efficiency, I am consolidating the installation of all
the hadoop daemons into one uber hadoop-base container.

This prevents unnecessary duplication of package downloads and keeps my
image inheritance tree smaller.

I also needed this to create a "client" container that the user would
run when wanting to interact with any of the daemons directly.